[UPDATED] total field type of transaction repo

+++ backEnd/src/main/java/com/happy/shoppingcart/api/service/TransactionService.java
@@ -19,7 +19,7 @@ public class TransactionService {
         Transaction res = result.get();
         TransactionGetResponse transactionGetResponse = new TransactionGetResponse(
                 res.getPoint(),
-                res.getTotal(),
+                res.getTotal().doubleValue(),
                 res.getShippingFee()
         );
         return transactionGetResponse;
